refactor: move deploy constants in another module
